+++ gui_engine/engine/acc_sw_cover.c
@@ -36,7 +36,7 @@ void cover_blit_2_rgb565(draw_img_t *image, struct gui_dispdev *dc,
     {
         return;
     }
-    uint32_t image_off = sizeof(struct gui_rgb_data_head) + (uint32_t)(image->data);
+    uint32_t image_off = sizeof(struct gui_rgb_data_head) + (uint32_t)(uintptr_t)(image->data);
     struct gui_rgb_data_head *head = image->data;
     char img_type = head->type;
 
@@ -56,7 +56,7 @@ void cover_blit_2_rgb565(draw_img_t *image, struct gui_dispdev *dc,
             for (uint32_t j = x_start; j <= x_end; j++)
             {
 
-                uint16_t pixel = (*((uint16_t *)read_off + j));
+                uint16_t pixel = (*((uint16_t *)(uintptr_t)read_off + j));
                 writebuf[write_off + j] = pixel;
             }
         }
