+++ gui_engine/engine/acc_sw_raster.c
@@ -71,7 +71,7 @@ static void do_raster_pixel(uint8_t *writebuf, int write_off, int image_base, ui
 
     if (input_type == RGB565)
     {
-        color_rgb565_t *pixel = (color_rgb565_t *)image_base + image_off;
+        color_rgb565_t *pixel = (color_rgb565_t *)(uintptr_t)image_base + image_off;
         source_alpha = 0xff;
         source_red = pixel->r << 3;
         source_green = pixel->g << 2;
@@ -79,7 +79,7 @@ static void do_raster_pixel(uint8_t *writebuf, int write_off, int image_base, ui
     }
     else if (input_type == RGB888)
     {
-        color_rgb888_t *pixel = (color_rgb888_t *)image_base + image_off;
+        color_rgb888_t *pixel = (color_rgb888_t *)(uintptr_t)image_base + image_off;
         source_alpha = 0xff;
         source_red = pixel->r;
         source_green = pixel->g;
@@ -87,7 +87,7 @@ static void do_raster_pixel(uint8_t *writebuf, int write_off, int image_base, ui
     }
     else if (input_type == ARGB8565)
     {
-        color_argb8565_t *pixel = (color_argb8565_t *)image_base + image_off;
+        color_argb8565_t *pixel = (color_argb8565_t *)(uintptr_t)image_base + image_off;
         source_alpha = pixel->a;
         source_red = pixel->r << 3;
         source_green = pixel->g << 2;
@@ -95,7 +95,7 @@ static void do_raster_pixel(uint8_t *writebuf, int write_off, int image_base, ui
     }
     else if (input_type == RGBA8888)
     {
-        color_argb8888_t *pixel = (color_argb8888_t *)image_base + image_off;
+        color_argb8888_t *pixel = (color_argb8888_t *)(uintptr_t)image_base + image_off;
         source_alpha = pixel->a;
         source_red = pixel->r;
         source_green = pixel->g;
@@ -105,7 +105,7 @@ static void do_raster_pixel(uint8_t *writebuf, int write_off, int image_base, ui
 
     if (dc_bytes_per_pixel == 2)
     {
-        color_rgb565_t *pixel = (color_rgb565_t *)writebuf + write_off;
+        color_rgb565_t *pixel = (color_rgb565_t *)(uintptr_t)writebuf + write_off;
         target_alpha = 0xff;
         target_red = pixel->r << 3;
         target_green = pixel->g << 2;
@@ -113,7 +113,7 @@ static void do_raster_pixel(uint8_t *writebuf, int write_off, int image_base, ui
     }
     else if (dc_bytes_per_pixel == 3)
     {
-        color_rgb888_t *pixel = (color_rgb888_t *)writebuf + write_off;
+        color_rgb888_t *pixel = (color_rgb888_t *)(uintptr_t)writebuf + write_off;
         target_alpha = 0xff;
         target_red = pixel->r;
         target_green = pixel->g;
@@ -121,7 +121,7 @@ static void do_raster_pixel(uint8_t *writebuf, int write_off, int image_base, ui
     }
     else if (dc_bytes_per_pixel == 4)
     {
-        color_argb8888_t *pixel = (color_argb8888_t *)writebuf + write_off;
+        color_argb8888_t *pixel = (color_argb8888_t *)(uintptr_t)writebuf + write_off;
         target_alpha = pixel->a;
         target_red = pixel->r;
         target_green = pixel->g;
@@ -216,7 +216,7 @@ void do_raster_no_rle(draw_img_t *image, struct gui_dispdev *dc, gui_rect_t *rec
         return;
     }
 
-    uint32_t image_base = sizeof(gui_rgb_data_head_t) + (uint32_t)(image->data);
+    uint32_t image_base = sizeof(gui_rgb_data_head_t) + (uint32_t)(uintptr_t)(image->data);
     gui_rgb_data_head_t *head = image->data;
     char input_type = head->type;
     int16_t source_w = image->img_w;
@@ -270,14 +270,14 @@ static void get_rle_pixel(draw_img_t *image, int x, int y, uint8_t *pixel)
 {
     gui_img_file_t *file = (gui_img_file_t *)image->data;
     imdc_file_t *compressed = (imdc_file_t *)(&(file->data.idc_file));
-    uint32_t line = (uint32_t)compressed + compressed->compressed_addr[y];
+    uint32_t line = (uint32_t)(uintptr_t)compressed + compressed->compressed_addr[y];
     int location = 0;
     imdc_rgb565_node_t *node = NULL;
 
 
     do
     {
-        node = (imdc_rgb565_node_t *)line;
+        node = (imdc_rgb565_node_t *)(uintptr_t)line;
         location += node->len;
         line = line + sizeof(imdc_rgb565_node_t);
     }
@@ -333,7 +333,7 @@ void do_raster_use_rle(draw_img_t *image, struct gui_dispdev *dc, gui_rect_t *re
             uint8_t rle_pixel[4];
             get_rle_pixel(image, x, y, rle_pixel);
 
-            do_raster_pixel(writebuf, write_off, (int)rle_pixel, 0, input_type, dc_bytes_per_pixel,
+            do_raster_pixel(writebuf, write_off, (int)(uintptr_t)rle_pixel, 0, input_type, dc_bytes_per_pixel,
                             opacity_value,
                             blend_mode);
         }
